+++ integrator/simpson_integrator.hpp
@@ -26,7 +26,7 @@ public:
 	//! Конструктор, принимает в качестве параметра число N точек разбиения
 	/** \param h шаг разбиения - параметр, определяющий точность алгоритма (меньше h - выше точность)
 	*/
-	simpson_integrator (double h)
+	simpson_integrator (long double h)
 		: h(h)
 	{ }
 
@@ -38,15 +38,15 @@ public:
 	}
 
 	//! Выполняет численное интегрирование и возвращает результат
-	T integrate (boost::function < T(double) > f, double x0, double x1) {
+	T integrate (boost::function < T(long double) > f, long double x0, long double x1) {
 		int N = int ((x1 - x0) / h);
 		N = std::max (N, 10);
 		if (N % 2)  ++N;
 
-		double h = (x1 - x0) / N;
+		long double h = (x1 - x0) / N;
 		T res = T();
 		for (int i=0; i<=N; ++i) {
-			double x = x0 + h * i;
+			long double x = x0 + h * i;
 			res += f(x) * ((i==0 || i==N) ? 1 : (i%2==0) ? 2 : 4);
 		}
 		return res * (x1 - x0) / N / 3.0;
@@ -55,7 +55,7 @@ public:
 protected:
 
 	//! параметр метода - h - шаг интегрирования
-	double h;
+	long double h;
 
 }; // class simpson_integrator
